改为protobuff来实现协议

+++ protoqueue-common/src/main/java/net/protoqueue/rpc/Parameter.kt
@@ -13,4 +13,4 @@ class RequestParameter(val timeout: Int = 10000)
 /**
  * 返回参数
  */
-class ResponseParameter(val resCode: Long?, val resMsg: String?)
+class ResponseParameter(val resCode: Int?, val resMsg: String?)
